Add discussion for day 17

+++ src/day17/activity/Bottle.java
@@ -7,7 +7,7 @@ public class Bottle {
     String brand;
     String color;
     String type;
-    double capacity;
+    int capacity;
     double weightInGrams;
     
     // Purpose / method
